Revise `EncodedTokenEvent` type.
